Update eloquent.md
Hi there, this is my first "fork" on github, so I hope I am doing it right :)

In the Laravel 3 docs for Eager Loading (laravel.com/docs/database/eloquent/#eager) there was a short section about using $include in a model in order to avoid having to use with() all the time.

I looked for this in the Laravel 4 docs (http://four.laravel.com/docs/eloquent#eager-loading), but could not find it. From the source code, I found that the variable in the model is now called $with (rather than $include). So I thought it would be a good idea to add it (back) to the documentation for Laravel 4.