Improving Schema docs
Copy of https://github.com/laravel/laravel/issues/2066

So I've just spend some time setting up a database with migrate, for the first time. In the database I need to reference users alot, which has a primary key make with increments('id').

But trying to do artisan migrate, I got alot of errors. In the I've found somewhere that you need to use integer('user_ref')->unsigned(), when linking to increments values. Nothing of this shows in the docs. So I suggust this gets added to the docs, for future users.

PS. after fixing that error I loved the db creation. Alot better then making the tables by hand :)