Updates to View::composer()
It's a common thing to bind the `layout` view instead of giving an array of child views. It's a better usage on most cases.

For example, if `dashboard, profile, update, messages, wall` etc. sub views are all extending the `layouts.account_layout` layout, you can pass `layouts.account_layout` to composer instead of all the pages.

Page specific things has to be binded using `with` method, so it fits the logic of using View Composers.