Specifying twitter bootstrap version compatiblity
Twitter bootstrap 3.0 has been released. For the pagination the class "pagination" should be applied to the `<ul>` not the `<div>` so the views are no longer compatible out of the box with Twitter Bootstrap.

But I don't feel like this is worded well as it's compatible with version 1.0-2.3 anyone want to suggest a further edit?