No need to call ->get()
Calling ->get() on belongsToMany relations when calling any form of has returns an empty set of results without error. Removing the ->get() will always return the correct call. Someone correct me if I am wrong but this seems like a bug in the code and was a source of huge headache.