Resource routes after other routes
Resource routes have to be declared after any other routes applying to the same path or a subpath of it. This behavior doesn't seem to be documented anywhere. I'm not sure if this is how it's supposed to work, but if so, I've written a brief explanation with an example.