Added note about relationship column when eager loading with callback
With many ORM's eager loading is done with a join so doesn't necessarily require the columns for the foreign key relationship in the select statement.  Due to the efficient method Eloquent uses, it does require that that particular column be included, which might not be immediately obvious.

Alternatively to modifying the doc, the getEager method could check the eager query columns to see if it's a wildcard or the particular foreign key relation is included, and if not, automatically add it to the query if not.