Specifying 'gulpfile.js' as the file to edit.
It may seem obvious to some, but for people who are new to Gulp or unclear on what Elixir actually does, the Usage section should clearly state that "gulpfile.js" is the file that you edit when using the examples listed.  The file "gulpfile.js" wasn't previously mentioned by name anywhere in the Elixer docs page.