Standarize indentation to tabs instead of mixture.
Currently the indentation is, for the most part, tabs. However, some of the
indentation is spaces, and there doesn't seem to be a good reason for a
lot of this indentation switch.

For example, some of the indentation is only 3 spaces or 7 spaces instead of
4 or 8. This is even more inconsistent and has probably happened over time
as a result of PRs.  This fixes inconsistencies such as that and makes
it 1 tab or 2 tabs.

This does not change specifically-spaced formatting to tabs, such as
aligning far-spaced -> in the queries markdown document and other
documents, meaning that ->'s are still perfectly aligned.

For example, the following is changed:

Old:

    if (Auth::attempt(['email' => $email, 'password' => $password, 'active' => 1]))
    {
        // The user is active, not suspended, and exists.
    }

New:

	if (Auth::attempt(['email' => $email, 'password' => $password, 'active' => 1]))
	{
		// The user is active, not suspended, and exists.
	}

Alternatively, the following is not changed:

	DB::table('users')
	        ->join('contacts', function($join)
	        {
	        	$join->on('users.id', '=', 'contacts.user_id')
	        	     ->where('contacts.user_id', '>', 5);
	        })
	        ->get();

This is to preserve the formatting of -> alignment.