Custom Error Messages Method
A new sub-section was added to Form Request Validation. This subsection is supposed to be a reference only to show it is possible to override the error messages in a custom Request class. A link is provided to the Custom Error Messages section to give a quick link to more information and examples.