Eloquent's vs. Collections's lists() correction.
Eloquent and Collection each have their own distinct lists() method, and both were changed to return a Collection instead of an array.  The upgrade guide failed to mention that Collection's lists() method now returns a Collection, but it did mention that Eloquent's lists() returns a Collection... but it mentioned that under the "Collection Class" section of the instructions.  Hopefully this will clear up the confusion.