Adding consistency in method injection documentation
If you look at the section on accessing requests in controllers here:

http://laravel.com/docs/5.1/requests#accessing-the-request

And then look at the similar example on method injection in controllers here:

http://laravel.com/docs/5.1/controllers#dependency-injection-and-controllers

You'll see a small discrepancy: the "Request" section gives an example of what the route looks like, while the controller section does not. This PR simply makes the "Controller" section consistent with the "Request" section, and adds the missing route example.