Restore useful content eliminated in 5.1
In 5.1, the `URLs To Controller Actions` section of this page was eliminated, and the explanation of `action()` was jammed into the `Naming Controller Routes` section. This was a step backwards in that the previous section name was more appropriate, and two important methods were scrubbed from the page.

Mention of the `URL::setRootControllerNamespace()` method was eliminated. This is a hugely useful method that is probably under-utilized because it is little-known. My feeling is that it has a place on this page.

Similarly, the `Route::currentRouteAction()` method was eliminated. There is no clear explanation as to why (at least that I'm able to find).

This change restores the `URLs to Controller Actions` section, verbatim, as it appeared in 5.0.