Add sample function
When trying detecting exception by using try-catch statement, it failed
in case of setting “Exception $e” as a parameter. However, it works
case of setting “\Exception $e”. We think it isn’t easy to realize the
cause. So, we hope this sample function will be added as a example of
combination of try-catch and the logger.