Restore description of queue:work command
From 5.0 to 5.1 the description of the queue:work command was removed. I'm not sure whether this was intentional, since it was part of a much larger batch of edits (059a01b65089819c4d81b0378be694e2d5a1bfdd).

Right now if you read through the queue docs, it's not really clear what difference is between `queue:listen` and `queue:work` is - there is no description of the latter.