add note for creating multiple factory files
if you have a large number of models or many factory types for each model, having them all in the `ModelFactory.php` file quickly becomes unmanageable.  this note explains to users they are free to create multiple files in the `factories` directory, and they don't all have to go in the `ModelFactory.php` file.

IMHO I think multiple files should be the preferred and encouraged way. In fact, we could even make an Artisan `make` command for it as well. Wondering what your thoughts are on that?