Provider boot() method and RouteServiceProvider changes.
Upgrading to 5.3 was a bit more than what the upgrade guide detailed. I also had to copy over the providers in github as the boot definition changed and threw errors. As I was updating the RouteServiceProvider it setup new route locations to routes/web.php routes/api.php, so had to update that as well. Moving Http/routes.php to routes/web.php and creating a new routes/api.php from github default.