Add note to run view:clear
Not sure where you want to stick this, but it is pretty important to run `view:clear` to avoid all sorts of view errors once the 5.4 upgrade has been completed. I don't think it is common knowledge to run this command after an upgrade. I don't mind if you close this PR and add the info in yourself where appropriate.