Observers != Events
It appears, in my testing that if you're using the `$events` property to define "traditional" events then the Observable class will not fire. E.g.,

```php
class Document extends Model
{
    protected $events = [
        'saved' => DocumentSaved::class,
    ];

    public function boot() {
        static::observe(DocumentObserver::class);
    }
```

```php
class DocumentObserver
{

    public function saved(Document $document)
    {
        dd('does not fire');
    }
```

I uncovered this while simplifying some code. I was in the process of removing an event for an observable and found it it didn't fire until I removed the `$events` property. Not sure if this is intended but the documentation update should make it a little more clear for future people.