more clarity on CreateFreshApiToken Middleware
As per the documentation it says to define the middleware in the web groups which is applied to all the routes by default.
If the all the routes are defined in the web.php file than this adds the `laravel_token` to all the requests regardlesss of whatever the guards logged in. Which causes a security flaw by which any authenticated user can issue/access routes of the passport guard applied.
I was having this issue which is a intended behaviour but it was nowhere mentioned in the documentation, So I though it might help for the users who has multiple auth guard setups.