Update upgrade.md
It's important to also check the update for the 3rdparty packages installed on Laravel.

I just wasted hours looking for what was avoiding whoops show the errors, and it actually was some kind of bug on DebugBar package. I updated it to the last version and it started working again.