Refactor method to checkout to the latest stable release
As pointed out in #3673 this removes the need of PRs that solely changes
the homestead tag version.

It will now dynamically determine the latest release tag.

This syntax works well because the release cycle of homestead is linear,
but in case that in the future parallel versions of homestead are
released (like v6.3.x and v7.0.x at the same time) we can specify a
pattern to scope them.

For example:

```
git checkout $(git tag --list v7.* | tail -1)
```