Replaced recommended plugin
The documentation of the recommended plug-in `vagrant-bindfs` states that it is necessary to make additional settings in the `VagrantFile` file.
This is a problem when updating `Homestead`, because every time you will need to re-add this setting, which you can just forget.
The proposed plugin `vagrant-winnfsd` does not require additional configuration settings. You just need to install it and specify parameter `type: "nfs"`. That's it.