remove PHP mail driver
the PHP mail driver is no longer used. It is simply an alias for the sendmail driver.

the 'mail' driver will continue to work (for now), we just shouldn't document it to help avoid confusion.

with the current wording, the program may think the native PHP `mail()` function is still being used, which it is not.