Explain vue assertions in components
A common use case would be to create a dusk component to test vue component. I had problems with it because I didn't understand why it's not working like in example. Problem was I didn't take in account that we are already in context of component and selector in third param of assertVue() would look for it inside this component DOM. So this is my suggestion how to describe how to use them together.