Improve note for onOneServer
The "command" used for generating the SHA1 to ensure a task is run on one server in a group, if an artisan command, will include the full path to the PHP executable. If there is a difference in the install location here, even if using a centralized memcached or redis server, the task will be run twice as a different SHA1 hash will be generated.