Added additional info on running NPM.
Vagrant can be stubborn depending on the operating system and various settings. On macOS, running NPM from within Homestead may not work as expected. That is not necessarily an issue with Laravel Mix or NPM. I found that running NPM from the host machine works. This commit adds info on that.