add note that API and frontend have to share the same TLD
To me, it was not obvious right away that this was a requirement.

Having read GitHub issues over and over again and after chatting in the Discord, it seems it was not quite clear amongst people there either.

Even though it makes complete sense that this is a requirement, I think noting it in the docs is important as I have spend days trying to educate myself on this issue on why it wasn't working, instead of learning how the foundation is working.

So I hope this helps people not make the same mistake as me 🤦🏼‍♂️