Remove mention of disabling view expiration check
The PR to allow disabling checks for expired views (#31206) was reverted because it didn't account for third party views. The docs still mentions the 'expires' option and this confused as I didn't see it make any difference. Digging into the view Complier code made me realize that this update must have been removed.