update `sortDesc` documentation
The wording is a little confusing because these 2 methods do not share the same signature. `sort` accepts an optional callback, while `sortDesc` accepts an optional set of options.  Added clarification you cannot pass a callback to `sortDesc`, and a suggestion for how to use `sort` to achieve the desired results.