Add APP_URL to Github Action example
## Description
This PR adds a default `APP_URL=http://127.0.0.1:8000` environment to the Github Action configuration example for running dusk + a short description that Github will not serve the application on `http://localhost` which is the default `APP_URL` for a clean Laravel installation

## Why
If one creates a completely fresh Laravel installation then `APP_URL=http://localhost` which works for many use cases. If one simply copies this Github Actions configuration example provide here, one would expect that it works out of the box, but this is not the case since Github Actions does not respond on `http://localhost` (I have no source confirming this except testing).

## With this PR
With this PR, people will be able to simply copy paste this example and start using Dusk tests on Github actions without any further configuration.

Setting a value for `APP_URL` will override the value provided in any environment files, but this is not an issue since the provided configuration will only respond on the provided address.