Factory states: Add tip for using dynamic data
This is a gotcha with the new factories in 8.x. When generating dynamic data within a state transformation, e.g.:

```
public function anonymous()
    {
        $uuid = Str::uuid();
        return $this->state([
            'name' => 'anonymous_user_' . $uuid,
            'email' => $uuid . '@anonymous.com',
            'is_anonymous' => true,
            'password' => Hash::make(Str::random(10)),
        ]);
    }
```

The method only runs once however many models you're creating. So generating 100 instances of that state will create 100 models each with the same uuid.

Using a callback fixes that.