Update upgrade.md
Hi,

During the upgrade process, we've followed the guide, however we removed the references to the `allOnQueue()` and `allOnConnection()` from our code, but it that stopped the chain to process the next joib in the chain. 

Using these methods with the helper `dispatch()` returned the previous - correct - behaviour.

I don't know if this is going to be the way in the future, but just suggesting a note in the docs to help other people and avoid problems.

Thanks!