Clarify required software for Homestead
Minor change to clarify the software required for Homestead to function. The existing text is somewhat ambiguous and can lead to thinking that all of those provided options are required to be installed when in reality only one of them needs to be. Thanks to @petro-lazurko for bringing this to my attention.