Update sail.md
Small correction on the stop procedure for containers running in detached mode. 
`down` (current docs) will stop **and** remove containers. The equivalent of `Control + C` would be `stop` which only stops (and doesn't remove). 

The issue with running `down` is if any pruning is performed, those database volumes are now unlinked to containers and the data will be lost.