Add clarification on save() and saveMany() relationship methods
I recently had an experience where I used saveMany to add multiple models to a relationship. Before calling the saveMany() method I altered some properties on the models but did not intend for these changes to be persisted as I only wanted the relationship to be updated. 

From the official documentation, I wrongly understood that the save and saveMany() methods when called only updates the many-to-many relationship and not the models passed to the method as well and worked similarly to the sync() method, with the only difference being saveMany() takes model instances as input instead of an array of model ids. 

I realize the methods do have the word "save" in them, but as a new Laravel developer relying heavily on the documentation, I feel some clarification here would not hurt.