Update authentication.md
Changed the manual authentication example:

Two parts:
• Validate the post data coming in exists, and the email is an email.
• Pass a translated error message (shipped standard in lang/en/auth.php)

I realise it's not required, but it shows some best practice perhaps for other routes - we don't assume data exists, and blindly pass it to a function. The translation part introduces how translations can be used from the start in an application, very easily.