Wrong return value of sole funciton in example (#7098)
The current documentation for the `sole` method in Collections has an example which suggests the expected result. However, the expected result seems to be wrong. Instead of (int) 3, I expect (int) 2 to be the result. Implementing the given example confirms this.
The current proposed change is simply to replace "3" by "2" in the documentation.