Include a Note that Breeze Should Be Run on New Application
I noted that running breeze install command on an existing application overwrites the existing files which will be quite undesirable if the developer doesn't have any backup. I personally could have lost some progress on the project because I was not careful enough to read the instruction in the docs that it should be run in the documentation and the documentation didn't emphasize this limitation enough. However, I have seen that the overwrites aren't big enough but, without prior backup could lead to losing serious progress.