Add hint about performance when using whereDate etc.
Unfortunately I see these methods a lot in Laravel tutorials and in real project. People dont know what they are doing, but think this is just a fancy way to query the database.

But if you depend on indexes for database performance, because there is a signifiat amount of data and a significant load, these methods can make the query really slow and even break the application.