Change condition query example code
The conditional query example code includes a second param for the callback, but this param does not appear to be used.  Instead it uses the `$value` used for the original condition.

I found this confusing when I tried to implement a conditional query with an `isset` condition, and then used the variable from the `isset` check as the second param.  It was not performing as expected because the callback is using the `$value` not the provided variable.

The proposed change shifts the callback's second param to a `use` statement for the callback, making sure that the variable is unchanged when used in the callback.