Use `sail stop` command instead of `sail down` command
The current documentation recommends the use of `sail down` command to stop the containers execution. This command stops and also removes the containers in Docker. A better option would be to use `sail stop` instead which only stops the containers but doesn't remove them.

This is better for using services like Vapor. If you remove the containers next time you need to authenticate again with Vapor before being able to deploy.