Fix documentation for collections `duplicate` methods
The documentation for Eloquent Collections implies that all base Collections methods work on Eloquent Collections too. As noted in https://github.com/laravel/framework/issues/34416, this is not actually true for the `duplicates()` and `duplicatesStrict()` methods.

This PR adds notes to those two methods to indicate that they perform differently for Eloquent Collections.