Instruct users to change image name when customizing Sail
If users are developing multiple apps on the same machine and have customized the Dockerfile for different apps in different ways (or customized one but not another), Docker may use the image from one application to create a container for a different application, when in reality there should be two unique images. Sail is a great tool for users who are just becoming familiar with Docker, and I think this explicit instruction may save some headaches (it would have for me, anyway).