Added note to upsert method about unique index
Hi,

I was using the `upsert` method provided by eloquent and it was still inserting rows. i checked through google etc and found out that you must have `primary or unique` index on uniqueBy columns but this isn't documented in the docs.

this pr adds it and i checked laravel database grammer's classes and `compileUpsert()` method have `uniqueBy` argument in all of them.