Update wording explaining after _commit
The section explaining the after_commit queue option was a bit unclear with regards to what transaction(s) Laravel waits for before dispatching a job within a database (see StackOverflow post asking this exact question - https://stackoverflow.com/questions/67943199/clarification-for-aftercommit-in-laravel-queues).

It is also a bit unclear what happens in cases where nested database transactions are used, for instance, a nested transaction where the parent transaction fails, but the child transaction succeeds eg:

`DB::BeginTransaction():
      //logic

      DB::BeginTransaction():
         //dispatch job here
      DB::commit();

      //more logic but an error occurs here
DB::commit()`

In the example above, the job should not be dispatched because the parent transaction failed, even though the child transaction completed successfully. The proposed changes would make this, in my opinion, a bit clearer for newcomers to Laravel.