Update starter-kits.md to mention Fortify
Fortify can sometimes get lost in the documentation. I’ve seen many articles and stories mentioning the starter-kits and forum posts with people asking how to customise or remove X from the frontend. I myself have used a starter-kit only to remove the frontend in the past. Like how the Fortify docs reference the starter-kits, to starter-kits should reference Fortify as an alternative.