Updating custom throttle response. (#7950)
* Updating custom throttle reponse.

Added the 2 parameters that the custom response callback is called with in the ThrottleRequests::buildException method. I tried to implement my own custom response and lost the throttle headers (X-RateLimit-Limit, etc). It was not obvious to me how to get the header info into my response. I tried to dig through the code, but missed the spot where the callback is called. Got a response to my problem on Laracasts, through I'd help put it in the docs for others.

* Update routing.md

Co-authored-by: Taylor Otwell <taylor@laravel.com>