Notice `dispatchAfterResponse` needs webserver with FastCGI (#8237)
Just like in [terminable middleware](https://laravel.com/docs/6.x/middleware#terminable-middleware), the user should be noticed that the web server should implement FastCGI.

I ran through this issue myself on a dead simple app. I decided to go with a  database driver queue, but it would be a better experience if I had known faster, it took me a while and a chat with a friend to realize this :)

I hope I made the docs better! Thanks for Laravel ❤️