add clarification for job attempts
I stumbled on this edge case (of feature) in laravel while working on a client code base and couldn't find any documentation for it. 

Apparently a job with `0` $tries will be retried indefinitely. What is more, I tried defining `0` on the job and `1` on the worker, because if `0` meant infinite then,  `1` attempt  would be more specific to `infinite` attempts right? But the laravel workers doesn't check for specificity, it is the default if the Job didn't specify anything. 

The word `specific` threw me off, because what does it mean to be more specific? In this case it meant: The job as a `$tries` set, which means it is more granular and thus more specific. While I read it as, a `$tries` value which is more specific, and thus the least amount of `attempts` .