import correct Builder namespace
when constraining the eager loaded relationship, the class actually passed into the closure is an instance of the actual relationship class, which implements `Illuminate\Contracts\Database\Eloquent\Builder`, and **not** `Illuminate\Database\Eloquent\Builder`.

it would be very easy for a user to be confused, since in most of these other examples an `Illuminate\Database\Eloquent\Builder` is actually passed to the closure.