Remove duplicate service provider registration
Hi guys, I just noticed there is a duplicate in the AppServiceProvider register method for the Telescope Service Provider. As I know, there is just one Telescope Service Provider, but there are two ways how to resolve the class (either by passing the full path into the closure or by just entering the classname and make use of the "use" statement). So I guess what happened here is that both ways have been added to the doc, but it does confuse, because as per current documentation the service provider would be registered twice? .. Anyways, please let me know if I miss something.