Update ISO-8601 date example in Eloquent: Mutators & Casting
This commit replaces the actual date example in the "Eloquent: Mutators & Casting" section with a more generic representation of the UTC ISO-8601 format. The change aims to prevent search engines from misinterpreting the example date as the publication date of the documentation page and to provide a clearer representation of the format for readers.